requirements.txt
fastapi===0.115.0
python-dotenv
pytest==6.2.4
uvicorn
starlette
starlette-context
requests
beautifulsoup4